Rename fr_QC to fr_CA
